Proyecto Final

+++ Calendario-FrontEnd/src/components/ui/DeleteEventFab.js
@@ -1,14 +1,42 @@
 import React from 'react'
 import { useDispatch } from 'react-redux'
 import { eventStartDelete } from '../../actions/events';
+import Swal from "sweetalert2";
 
 export const DeleteEventFab = () => {
 
   const dispatch = useDispatch();
 
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger'
+    },
+    buttonsStyling: false
+  })
+
   const handleDelete = () => {
-    dispatch(eventStartDelete());
+    swalWithBootstrapButtons.fire({
+      title: '¿Estas seguro?',
+      text: "No se podra revertir",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Si, eliminar!',
+      cancelButtonText: 'No, cancelar!',
+      reverseButtons: true
+    }).then((result) => {
+      //si el usuario confirme el borrado, se borrara el evento, si no no
+      if (result.isConfirmed) {
+        dispatch(eventStartDelete());
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        swalWithBootstrapButtons.fire('Cancelado','Se ha cancelado el borrado','error')
+      }
+    }
+      )
   }
+
+    
+    
 
   return (
     <button
